+++ src/main/java/com/motaharinia/msutility/custom/customjson/serializer/CustomEnum.java
@@ -1,0 +1,5 @@
+package com.motaharinia.msutility.custom.customjson.serializer;
+
+public interface CustomEnum {
+    String getValue();
+}