+++ src/test/java/com/motaharinia/msutility/custom/customjson/JsonUnitTest.java
@@ -14,8 +14,6 @@ import org.springframework.context.support.ReloadableResourceBundleMessageSource
 
 import java.time.Instant;
 import java.util.*;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.fail;
@@ -27,7 +25,7 @@ import static org.assertj.core.api.Assertions.fail;
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 class JsonUnitTest {
 
-    private CustomObjectMapper mapper =null;
+    private CustomObjectMapper mapper = null;
 
     /**
      * این متد مقادیر پیش فرض قبل از هر تست این کلاس تست را مقداردهی اولیه میکند
@@ -35,15 +33,15 @@ class JsonUnitTest {
     @BeforeEach
     void beforeEach() {
         Locale.setDefault(new Locale("fa", "IR"));
-        mapper= new CustomObjectMapper(getMessageSource());
+        mapper = new CustomObjectMapper(getMessageSource());
         //این قطعه تنظیمات در پروژه های دیگر زمان بالا آمدن context از فایلهای properties میتواند خوانده  و تنظیم شود
-        mapper.configure(SerializationFeature.INDENT_OUTPUT,true);
-        mapper.configure(SerializationFeature.FAIL_ON_EMPTY_BEANS,false);
-        mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES,false);
-        mapper.configure(DeserializationFeature.FAIL_ON_IGNORED_PROPERTIES,false);
-        mapper.configure(DeserializationFeature.ACCEPT_SINGLE_VALUE_AS_ARRAY,true);
-        mapper.configure(DeserializationFeature.ACCEPT_EMPTY_STRING_AS_NULL_OBJECT,true);
-        mapper.configure(DeserializationFeature.ACCEPT_EMPTY_ARRAY_AS_NULL_OBJECT,true);
+        mapper.configure(SerializationFeature.INDENT_OUTPUT, true);
+        mapper.configure(SerializationFeature.FAIL_ON_EMPTY_BEANS, false);
+        mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+        mapper.configure(DeserializationFeature.FAIL_ON_IGNORED_PROPERTIES, false);
+        mapper.configure(DeserializationFeature.ACCEPT_SINGLE_VALUE_AS_ARRAY, true);
+        mapper.configure(DeserializationFeature.ACCEPT_EMPTY_STRING_AS_NULL_OBJECT, true);
+        mapper.configure(DeserializationFeature.ACCEPT_EMPTY_ARRAY_AS_NULL_OBJECT, true);
     }
 
     /**
@@ -55,9 +53,9 @@ class JsonUnitTest {
     }
 
 
-    private ReloadableResourceBundleMessageSource getMessageSource(){
-        ReloadableResourceBundleMessageSource messageSource= new ReloadableResourceBundleMessageSource();
-        messageSource.setBasenames("classpath:lang/businessexception","classpath:lang/customvalidation","classpath:lang/usermessage","classpath:lang/comboitem");
+    private ReloadableResourceBundleMessageSource getMessageSource() {
+        ReloadableResourceBundleMessageSource messageSource = new ReloadableResourceBundleMessageSource();
+        messageSource.setBasenames("classpath:lang/businessexception", "classpath:lang/comboitem", "classpath:lang/customvalidation", "classpath:lang/exception", "classpath:lang/usermessage");
         messageSource.setCacheSeconds(5);
         messageSource.setDefaultEncoding("UTF-8");
         return messageSource;
@@ -74,25 +72,29 @@ class JsonUnitTest {
             Instant dateOfBirth = Instant.now();
 
             //مدل جاوا حاوی تاریخ میلادی برای ارسال به کلاینت
-            JsonDto jsonDto = new JsonDto(new CustomDate(dbDate), Stream.of(EtcItemGender.values()).map(EtcItemGender::getValue).collect(Collectors.toList()),dateOfBirth.toEpochMilli());
-            ClientResponseDto<JsonDto> dto = new ClientResponseDto<>(jsonDto,"USER_MESSAGE.FORM_SUBMIT_SUCCESS");
+            JsonDto jsonDto = new JsonDto(new CustomDate(dbDate), EtcItemGender.GENDER_FEMALE, dateOfBirth.toEpochMilli());
+            ClientResponseDto<JsonDto> dto = new ClientResponseDto<>(jsonDto, "USER_MESSAGE.FORM_SUBMIT_SUCCESS");
 
             //سریالایز مدل جاوا در کنترلر در زمان ارسال به کلاینت
-            String jsonString= this.mapper.writeValueAsString(dto);
+            String jsonString = this.mapper.writeValueAsString(dto);
+
+            System.out.println("jsonString:" + jsonString);
 
             //تستهای سریالایز
-            //تست عدم وجود خطا
+            //تست عدم وجود خطاs
             assertThat(jsonString).isNotNull();
             //تست تبدیل تاریخ میلادی به جلالی جهت ارسال به کلاینت
             //برای اینکه بتوانیم سال شمسی را در رشته جیسون پیدا کنیم این شیی را میسازیم
-            CustomDate jalaliCustomDate= CalendarTools.gregorianToJalaliCustomDate(dbDate);
+            CustomDate jalaliCustomDate = CalendarTools.gregorianToJalaliCustomDate(dbDate);
             assertThat(jsonString).contains(jalaliCustomDate.getYear().toString());
 
             //تست  رشته ترجمه "ای تی سی آیتم" به زبان فارسی
             assertThat(jsonString).contains("زن");
 
             //دیسریالایز رشته جیسون دریافتی از کلاینت به مدل در کنترلر
-            dto = this.mapper.readValue(jsonString, new TypeReference<ClientResponseDto<JsonDto>>() {
+            jsonString = jsonString.replace("زن", EtcItemGender.GENDER_FEMALE.toString());
+
+            dto = this.mapper.readValue(jsonString, new TypeReference<>() {
             });
 
             //تست های دیسریالایز
@@ -103,7 +105,7 @@ class JsonUnitTest {
             Calendar dbDateCalendar = Calendar.getInstance(TimeZone.getTimeZone("Asia/Tehran"));
             dbDateCalendar.setTime(dbDate);
             assertThat(dto.getData().getCustomDate().getYear()).isEqualTo(dbDateCalendar.get(Calendar.YEAR));
-            assertThat(dto.getData().getCustomDate().getMonth()).isEqualTo(dbDateCalendar.get(Calendar.MONTH)+1);
+            assertThat(dto.getData().getCustomDate().getMonth()).isEqualTo(dbDateCalendar.get(Calendar.MONTH) + 1);
             assertThat(dto.getData().getCustomDate().getDay()).isEqualTo(dbDateCalendar.get(Calendar.DAY_OF_MONTH));
             assertThat(Date.from(Instant.ofEpochMilli(dto.getData().getDateOfBirth()))).isEqualTo(Date.from(dateOfBirth));
 
@@ -118,7 +120,7 @@ class JsonUnitTest {
     void serializeEmptyDtoTest() {
         try {
             MembershipRequestFrontDtoUpdate membershipRequestFrontDtoUpdate = new MembershipRequestFrontDtoUpdate();
-            String jsonString= this.mapper.writeValueAsString(membershipRequestFrontDtoUpdate);
+            String jsonString = this.mapper.writeValueAsString(membershipRequestFrontDtoUpdate);
             assertThat(jsonString).isNotNull();
         } catch (Exception ex) {
             fail(ex.toString());